+++ software/ai/hl/stp/ui.cpp
@@ -161,8 +161,8 @@ void AI::HL::STP::draw_offense(const World &world, Cairo::RefPtr<Cairo::Context>
 		const double dx = (x2 - x1) / (grid_x + 1) / 2;
 		const double dy = (y2 - y1) / (grid_y + 1) / 2;
 
-		for (int i = 1; i <= 2 * grid_y + 1; i += 2) {
-			for (int j = (i / 2 + 1) % 2 + 1; j <= 2 * grid_x + 1; j += 2) {
+		for (unsigned int i = 1; i <= 2 * static_cast<unsigned int>(grid_y) + 1; i += 2) {
+			for (unsigned int j = (i / 2 + 1) % 2 + 1; j <= 2 * static_cast<unsigned int>(grid_x) + 1; j += 2) {
 				const double x = x1 + dx * j;
 				const double y = y1 + dy * i;
 				//const Point pos = Point(x, y);
